+++ .gitignore
@@ -1,0 +1,6 @@
+cy-test/node_modules/*
+conduit/README.md
+conduit/.settings*
+conduit/.project*
+conduit/.gitignore*
+conduit/.classpath*
