+++ cypress.sh
@@ -1,0 +1,4 @@
+cd cy-test
+npm install -g cypress
+npm install
+cypress run --spec "cypress/integration/**/*"
