+++ selenium.sh
@@ -1,0 +1,2 @@
+cd conduit/
+mvn clean test -DsuiteXmlFile=src/test/resources/testng.xml
